rewrote book.js
