+++ models/book.js
@@ -1,0 +1,29 @@
+const mongoose = require('mongoose');
+
+var BookSchema = new mongoose.Schema({
+  title: String,
+  author: String,
+  image: String,
+  releaseDate: String
+});
+
+var Book = mongoose.model('Book', BookSchema);
+
+// module.exports = Book;
+
+
+
+
+// // entire book.js so far
+// var mongoose = require('mongoose');
+
+// var BookSchema = new mongoose.Schema({
+//   title: String,
+//   author: String,
+//   image: String,
+//   releaseDate: String
+// });
+
+// var Book = mongoose.model('Book', BookSchema);
+
+// module.exports = Book;
