+++ models/index.js
@@ -6,5 +6,5 @@ mongoose.connection.openUri(process.env.DB_CONN);
 
 
 // add exports here
-module.exports.Book = require("./books.js");
+module.exports.Book = require("./book.js");
